% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgreview-package.R
\docType{package}
\name{pkgreview-package}
\alias{pkgreview}
\alias{pkgreview-package}
\title{pkgreview: Tools for Reviewing Other CRAN packages}
\description{
A single function that clones and opens another CRAN package in the IDE.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/pkgreview}
  \item \url{https://krlmlr.github.io/pkgreview}
  \item Report bugs at \url{https://github.com/krlmlr/pkgreview/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

}
\keyword{internal}
