% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revdep.R
\name{pkg_review}
\alias{pkg_review}
\title{Review a CRAN package}
\usage{
pkg_review(pkg, pak = TRUE)
}
\arguments{
\item{pkg}{The name of the package to review.}

\item{pak}{Logical, whether to install the package dependencies using \code{\link[pak:pak]{pak::pak()}}.}
}
\description{
This function is used to review a CRAN package by cloning it from the CRAN repository,
rebasing it to the latest version if needed, opening RStudio IDE, and optionally
installing the package's dependencies using \code{pak}.
}
\details{
It is in a package because a standalone is not a good fit.
}
